Separate utility and callback functions
